+++ packages/vue-language-core/src/generators/template.ts
@@ -67,6 +67,7 @@ export function generate(
 	sourceLang: string,
 	templateAst: CompilerDOM.RootNode,
 	hasScriptSetupSlots: boolean,
+	sharedTypesImport: string,
 	cssScopedClasses: string[] = [],
 ) {
 
@@ -177,7 +178,7 @@ export function generate(
 			const varName = validTsVar.test(tagName) ? tagName : capitalize(camelize(tagName.replace(/:/g, '-')));
 
 			codes.push(
-				`& import('./__VLS_types.d.ts').WithComponent<'${varName}', typeof __VLS_components, `,
+				`& import('${sharedTypesImport}').WithComponent<'${varName}', typeof __VLS_components, `,
 				// order is important: https://github.com/johnsoncodehk/volar/issues/2010
 				`"${capitalize(camelize(tagName))}", `,
 				`"${camelize(tagName)}", `,
@@ -460,7 +461,7 @@ export function generate(
 			codes.push([leftExpressionText, 'template', leftExpressionRange.start, capabilitiesPresets.all]);
 			formatCodes.push(...createFormatCode(leftExpressionText, leftExpressionRange.start, formatBrackets.normal));
 		}
-		codes.push(`] of (await import('./__VLS_types.d.ts')).getVForSourceType`);
+		codes.push(`] of (await import('${sharedTypesImport}')).getVForSourceType`);
 		if (source.type === CompilerDOM.NodeTypes.SIMPLE_EXPRESSION) {
 			codes.push(
 				...createInterpolationCode(
@@ -559,7 +560,7 @@ export function generate(
 		}
 
 		codes.push(
-			`const ${var_functionalComponent} = (await import('./__VLS_types.d.ts')).asFunctionalComponent(`,
+			`const ${var_functionalComponent} = (await import('${sharedTypesImport}')).asFunctionalComponent(`,
 			`${var_originalComponent}, `,
 			`new ${var_originalComponent}({`,
 			...createPropsCode(node, props, 'slots'),
@@ -605,12 +606,12 @@ export function generate(
 			tagOffsets.length ? ['', 'template', tagOffsets[0] + tag.length, capabilitiesPresets.diagnosticOnly]
 				: dynamicTagExp ? ['', 'template', startTagOffset + tag.length, capabilitiesPresets.diagnosticOnly]
 					: '',
-			`, ...(await import('./__VLS_types.d.ts')).functionalComponentArgsRest(${var_functionalComponent}));\n`,
+			`, ...(await import('${sharedTypesImport}')).functionalComponentArgsRest(${var_functionalComponent}));\n`,
 		);
 
 		if (tag !== 'template' && tag !== 'slot') {
 			componentCtxVar = `__VLS_${elementIndex++}`;
-			codes.push(`const ${componentCtxVar} = (await import('./__VLS_types.d.ts')).pickFunctionalComponentCtx(${var_originalComponent}, ${var_componentInstance})!;\n`);
+			codes.push(`const ${componentCtxVar} = (await import('${sharedTypesImport}')).pickFunctionalComponentCtx(${var_originalComponent}, ${var_componentInstance})!;\n`);
 			parentEl = node;
 		}
 
@@ -649,7 +650,7 @@ export function generate(
 		if (vScope?.type === CompilerDOM.NodeTypes.DIRECTIVE && vScope.exp) {
 
 			const scopeVar = `__VLS_${elementIndex++}`;
-			const condition = `(await import('./__VLS_types.d.ts')).withScope(__VLS_ctx, ${scopeVar})`;
+			const condition = `(await import('${sharedTypesImport}')).withScope(__VLS_ctx, ${scopeVar})`;
 
 			codes.push(`const ${scopeVar} = `);
 			codes.push([
@@ -709,7 +710,7 @@ export function generate(
 							slotDir.exp.loc.start.offset,
 							capabilitiesPresets.all,
 						],
-						`] = (await import('./__VLS_types.d.ts')).getSlotParams(`,
+						`] = (await import('${sharedTypesImport}')).getSlotParams(`,
 					);
 				}
 				else {
@@ -721,7 +722,7 @@ export function generate(
 							slotDir.exp.loc.start.offset,
 							capabilitiesPresets.all,
 						],
-						` = (await import('./__VLS_types.d.ts')).getSlotParam(`,
+						` = (await import('${sharedTypesImport}')).getSlotParam(`,
 					);
 				}
 			}
@@ -794,7 +795,7 @@ export function generate(
 				const eventVar = `__VLS_${elementIndex++}`;
 				codes.push(
 					`let ${eventVar} = { '${prop.arg.loc.source}': `,
-					`(await import('./__VLS_types.d.ts')).pickEvent(${componentCtxVar}.emit!, '${prop.arg.loc.source}' as const, (await import('./__VLS_types.d.ts')).componentProps(${componentVar}, ${componentInstanceVar})`,
+					`(await import('${sharedTypesImport}')).pickEvent(${componentCtxVar}.emit!, '${prop.arg.loc.source}' as const, (await import('${sharedTypesImport}')).componentProps(${componentVar}, ${componentInstanceVar})`,
 					...createPropertyAccessCode([
 						camelize('on-' + prop.arg.loc.source), // onClickOutside
 						'template',
@@ -1357,7 +1358,7 @@ export function generate(
 						prop.loc.start.offset,
 						capabilitiesPresets.diagnosticOnly,
 					],
-					`(await import('./__VLS_types.d.ts')).directiveFunction(__VLS_ctx.`,
+					`(await import('${sharedTypesImport}')).directiveFunction(__VLS_ctx.`,
 					[
 						camelize('v-' + prop.name),
 						'template',