+++ packages/vue-language-core/src/index.ts
@@ -1,13 +1,14 @@
 export * from './generators/template';
 export * from './languageModule';
-export * as scriptRanges from './parsers/scriptRanges';
 export * from './parsers/scriptSetupRanges';
 export * from './plugins';
 export * from './sourceFile';
 export * from './types';
-export * as localTypes from './utils/localTypes';
 export * from './utils/ts';
 export * from './utils/parseSfc';
+
+export * as scriptRanges from './parsers/scriptRanges';
+export * as sharedTypes from './utils/directorySharedTypes';
 
 export * from '@volar/language-core';
 export * from '@volar/source-map';